Added some major changest to the Web presentation layer
