+++ PresentationLayer/PresentationLayer/FormStudents.cs
@@ -2,39 +2,47 @@
 using DataLayer.Models;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
 using System.Linq;
-using System.Web;
-using System.Web.UI;
-using System.Web.UI.WebControls;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
 
-namespace PresentationLayerWeb
+namespace PresentationLayer
 {
-    public partial class WebFormStudents : System.Web.UI.Page
+    public partial class FormStudents : Form
     {
         private StudentBusiness studentBusiness = new StudentBusiness();
 
         private decimal d = 0m;
         // vrednost prema kojoj će se prikazivati studenti sa prosečnom ocenom većom od "d"
 
-        protected void Page_Load(object sender, EventArgs e)
+        public FormStudents()
+        {
+            InitializeComponent();
+        }
+
+        private void FormStudents_Load(object sender, EventArgs e)
         {
             GetAllStudents();
         }
 
         private void GetAllStudents()
         {
-            ListBoxStudent.Items.Clear();
+            listBoxStudents.Items.Clear();
 
-            List<Student> listOfStudents = this.studentBusiness.GetAllStudents();
+            List<Student> listOfStudents;
 
-            if (TextBoxStudentFilterByAverageMark.Text == "")
+            if (textBoxStudentFilterByAverageMark.Text == "")
             {
                 listOfStudents = this.studentBusiness.GetAllStudents();
                 // Ako je tekstualno polje za prosečnu ocenu prazno, nema filtriranja
             }
             else
             {
-                d = Convert.ToDecimal(TextBoxStudentFilterByAverageMark.Text);
+                d = Convert.ToDecimal(textBoxStudentFilterByAverageMark.Text);
 
                 listOfStudents = this.studentBusiness.GetStudentsWithAverageLargerThan(d);
                 // Ako je uneta prosečna ocena za filtriranje, poziva se sa biznis sloja
@@ -48,28 +56,28 @@ namespace PresentationLayerWeb
 
             foreach (Student s in listOfStudents)
             {
-                ListBoxStudent.Items.Add(s.Id + ". " + s.Name + " – " + s.IndexNumber + " — " + s.AverageMark);
+                listBoxStudents.Items.Add(s.Id + ". " + s.Name + " – " + s.IndexNumber + " — " + s.AverageMark);
             }
         }
 
-        protected void ButtonStudentInsert_Click(object sender, EventArgs e)
+        private void buttonInsertStudent_Click(object sender, EventArgs e)
         {
             Student s = new Student();
-            s.Name = TextBoxStudentName.Text;
-            s.IndexNumber = TextBoxStudentIndexNumber.Text;
-            s.AverageMark = Convert.ToDecimal(TextBoxStudentAverageMark.Text);
+            s.Name = textBoxStudentName.Text;
+            s.IndexNumber = textBoxStudentIndexNumber.Text;
+            s.AverageMark = Convert.ToDecimal(textBoxStudentAverageMark.Text);
 
-            LabelMessages.Text = this.studentBusiness.InsertStudent(s);
+            labelMessages.Text = this.studentBusiness.InsertStudent(s);
 
             GetAllStudents();
-        }
 
+        }
 
         // Pošto je pri učitavanju forme polje za prosečnu ocenu prazno,
         // kreiramo posebnu metodu koja se poziva klikom na odgovarajući taster, 
         // a koja ponovo poziva metodu GetAllStudents() na Formi i prikazuje
         // studente koji imaju prosečnu ocenu veću od one zadate u tekstualnom polju.
-        protected void ButtonStudentRefreshListBoxStudent_Click(object sender, EventArgs e)
+        private void buttonStudentRefreshListBox_Click(object sender, EventArgs e)
         {
             GetAllStudents();
         }
